# crates/server/src/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: DumbMonit HTTP API
  version: "0.1"
  summary: Monitoring for homelabs and small teams — the API the web UI uses, open to scripts and assistants.
  description: |
    Everything the DumbMonit web interface does goes through this API. It is an
    **alpha**: routes and fields may still change between releases. The
    `X-DumbMonit-Api-Version` header, present on every `/api` response, says
    which revision of this document the server implements.

    ## Authentication

    Two ways in, accepted on every protected route:

    - **API token** — `Authorization: Bearer dmt_…`, created in Settings →
      *API & assistants*. Meant for scripts, dashboards and assistants. No
      cookie is involved, so no anti-CSRF header is needed.
    - **Session cookie** — `dumbmonit_session`, what the web UI uses after
      `POST /api/auth/login`. Every state-changing request (`POST`, `PUT`,
      `DELETE`) sent with the cookie must also carry
      `X-Requested-With: DumbMonit`, or it is refused with `403`.

    A token has a scope. **`read`** grants what a *viewer* account sees: every
    `GET`. **`write`** grants what an *administrator* does: `POST`, `PUT` and
    `DELETE` as well. Whatever its scope, a token can never manage accounts,
    sessions, two-factor authentication, SSO, API or agent tokens, nor export
    or restore a backup: those operations are marked `x-token-scope: none` and
...
    `read` one.

...
    `read`, `write` or `none` (session only).

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      tags: [Accounts]
      summary: Sign in with a password
      description: >-
        `204` with a `Set-Cookie` header carrying the session. When the account
        has two-factor authentication, `200` with a pending ticket to finish
        with `POST /api/auth/login/totp`. Rate limited after five failures
        (`429`, `Retry-After` doubling from 30 s to 5 min).
      security: []
      x-token-scope: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "204":
          description: Signed in; the session cookie is set.
          headers:
            Set-Cookie:
              description: "`dumbmonit_session`, HttpOnly, SameSite=Lax."
              schema: { type: string }
        "200":
          description: Two-factor authentication required.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TotpChallenge" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/auth/login/totp:
    post:
      operationId: loginTotp
      tags: [Accounts]
      summary: Finish a sign-in with a two-factor code
      security: []
      x-token-scope: public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TotpLoginRequest" }
      responses:
        "204":
          description: Signed in; the session cookie is set.
          headers:
            Set-Cookie:
              description: "`dumbmonit_session`, HttpOnly, SameSite=Lax."
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Accounts]
      summary: End the current session
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      responses:
        "204": { description: Signed out; the cookie is cleared. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/auth/password:
    post:
      operationId: changePassword
      tags: [Accounts]
      summary: Change the password of the signed-in account
      description: Signs out every other session of the account.
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordChange" }
      responses:
        "204": { description: Password changed. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/auth/totp/enroll:
    post:
      operationId: enrollTotp
      tags: [Accounts]
      summary: Propose a two-factor secret
      description: Nothing is enforced until the secret is confirmed with `POST /api/auth/totp/verify`.
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordConfirmation" }
      responses:
        "200":
          description: The proposed secret.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TotpEnrolment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/auth/audit:
    get:
      operationId: listAuditLog
      tags: [Accounts]
      summary: The latest security events
      description: Administrator session only.
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 200 }
      responses:
        "200":
          description: Events, most recent first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEntry" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/targets:
    get:
      operationId: listTargets
      tags: [Devices]
      summary: Every device
      x-token-scope: read
      responses:
        "200":
          description: Devices.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Target" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
    post:
      operationId: createTarget
      tags: [Devices]
      summary: Add a device
      description: Profile detection starts in the background for SNMP.
      x-token-scope: write
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TargetPayload" }
            example:
              name: Core switch
              address: 192.168.1.10
              kind: snmp
              interval_secs: 60
              tags: {}
              credential: { type: snmp_community, community: public }
      responses:
        "201":
          description: The device.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Target" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/alerts:
    get:
      operationId: listActiveAlerts
      tags: [Alerts]
      summary: Active alerts
      description: >-
        Pending, firing, suppressed and recently resolved. `phase` is the
        engine's state; `effective_phase` is what to show (`suppressed` when a
        parent device is unreachable, named by `suppressed_by`). Alerts of
        deleted or paused devices are never listed.
      x-token-scope: read
      responses:
        "200":
          description: Alerts.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActiveAlert" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/backup/restore:
    post:
      operationId: restoreBackup
      tags: [Backup]
      summary: Restore a backup bundle (dry run unless `apply`)
      security: [{ cookieAuth: [] }]
      x-token-scope: none
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BackupRestoreRequest" }
      responses:
        "200":
          description: Per-section report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestoreReport" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/mcp:
    get:
      operationId: describeMcpEndpoint
      tags: [Assistants]
      summary: Explains the MCP endpoint (405)
      description: There is no server-sent stream; the body only says what this endpoint is.
      security: []
      x-token-scope: public
      responses:
        "405":
          description: Method not allowed; `Allow` names `POST`.
          headers:
            Allow:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/McpEndpointInfo" }
    delete:
      operationId: deleteMcpSession
      tags: [Assistants]
      summary: Session termination is not supported (405)
      description: The server is stateless and issues no `Mcp-Session-Id`, so there is nothing to terminate.
      security: []
      x-token-scope: public
      responses:
        "405":
          description: Method not allowed; `Allow` names `POST`.
          headers:
            Allow:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/McpEndpointInfo" }
    post:
      operationId: callMcp
      tags: [Assistants]
      summary: Model Context Protocol server (Streamable HTTP, JSON-RPC 2.0)
      description: |
        One JSON-RPC 2.0 request or notification per `POST`, answered in
        `application/json` (no SSE stream, no session).

        Protocol revisions:

        - **2026-07-28** (stateless): every request carries
          `params._meta["io.modelcontextprotocol/protocolVersion"]` and
          `clientCapabilities`, plus the `MCP-Protocol-Version`, `Mcp-Method`
          and (for `tools/call`) `Mcp-Name` headers, which must match the body
          (`400` with error `-32020` otherwise). `server/discover` lists the
          supported versions; an unsupported one gets `400` with error
          `-32022`; an unknown method `404` with `-32601`.
        - **2025-11-25**, **2025-06-18**, **2025-03-26**: the `initialize`
          handshake, then `tools/list` and `tools/call`; `ping` is answered.

        A notification is acknowledged with `202` and no body. The API token's
        scope applies tool by tool: a `read` token can call every read tool and
        gets an explanatory tool error (`isError: true`) from a write tool. The
        27 tools are documented at
        <https://dumbmonit.readthedocs.io/en/latest/using/assistant/>.
      x-token-scope: read
      parameters:
        - name: MCP-Protocol-Version
          in: header
          schema: { type: string, examples: ["2026-07-28"] }
        - name: Mcp-Method
          in: header
          schema: { type: string }
        - name: Mcp-Name
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JsonRpcRequest" }
            example:
              jsonrpc: "2.0"
              id: 1
              method: tools/call
              params:
                name: get_status
                arguments: {}
                _meta:
                  io.modelcontextprotocol/protocolVersion: "2026-07-28"
                  io.modelcontextprotocol/clientCapabilities: {}
      responses:
        "200":
          description: A JSON-RPC response (result or error).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonRpcResponse" }
        "202": { description: 'Notification accepted, no body.' }
        "400":
          description: Malformed request, header mismatch or unsupported protocol version (JSON-RPC error body).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonRpcResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Unknown method on a 2026-07-28 request (JSON-RPC error `-32601`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JsonRpcResponse" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ------------------------------------------- compatibilité Prometheus (hors /api)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: dmt_…
      description: >-
        API token created in Settings → API & assistants. `read` scope: every
        `GET`; `write` scope: also `POST`, `PUT`, `DELETE`. Never valid on
        operations marked `x-token-scope: none`. 120 calls per minute per token.
    cookieAuth:
      type: apiKey
      in: cookie
      name: dumbmonit_session
      description: >-
        Session opened by `POST /api/auth/login`. State-changing requests must
        also carry `X-Requested-With: DumbMonit`.
...
  headers:
    RetryAfter:
      description: Seconds to wait before trying again.
      schema: { type: integer }
    RateLimitLimit:
      description: Calls allowed per minute for this token.
      schema: { type: integer }
    RateLimitRemaining:
      description: Calls left in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until the window resets.
      schema: { type: integer }
    WwwAuthenticate:
      description: "`Bearer` (with `error=\"invalid_token\"` for an expired token)."
      schema: { type: string }
  responses:
    BadRequest:
      description: The request is malformed or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: No valid session or API token (unknown, revoked or expired token).
      headers:
        WWW-Authenticate: { $ref: "#/components/headers/WwwAuthenticate" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: >-
        Not allowed: viewer or `read` token on a write operation, token on a
        session-only operation or from a network it is not allowed from,
        missing `X-Requested-With: DumbMonit` on a cookie write.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: Not possible in the current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: Rate limited.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
        RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
        RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
        RateLimit-Reset: { $ref: "#/components/headers/RateLimitReset" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable message. }
...
    LoginRequest:
      type: object
      required: [password]
      properties:
        username: { type: string }
        password: { type: string, writeOnly: true }
    TotpChallenge:
      type: object
      required: [totp_required, pending]
      properties:
        totp_required: { type: boolean, const: true }
        pending: { type: string, description: 'Ticket valid five minutes, dead after five wrong codes.' }
    TotpLoginRequest:
      type: object
      required: [pending, code]
      properties:
        pending: { type: string }
        code: { type: string, description: A six-digit code or a recovery code. }
    PasswordChange:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: { type: string, writeOnly: true }
        new_password: { type: string, writeOnly: true }
    PasswordConfirmation:
      type: object
      required: [password]
      properties:
        password: { type: string, writeOnly: true }
...
    TotpEnrolment:
      type: object
      required: [secret, otpauth_uri, issuer, account]
      properties:
        secret: { type: string, description: Base32. }
        otpauth_uri: { type: string }
        issuer: { type: string }
        account: { type: string }
...
    AuditEntry:
      type: object
      required: [id, at, action]
      properties:
        id: { type: integer, format: int64 }
        at: { type: string }
        actor: { type: [string, "null"] }
        action: { type: string, examples: [login, login.failed, token.created, user.updated] }
        subject: { type: [string, "null"] }
        ip: { type: [string, "null"] }
...
    Target:
      type: object
      required: [id, name, address, kind, interval_secs, enabled, tags, credential_kind]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        address: { type: string }
        kind: { type: string, description: "A kind from `GET /api/collectors` (`snmp`, `proxmox`, `agent`, `http`, `pack.<id>`…)." }
        profile_id: { type: [string, "null"] }
        parent_id: { type: [integer, "null"], format: int64 }
        via_agent: { type: [integer, "null"], format: int64, description: Relay agent probing this device; `null` means the server. }
        interval_secs: { type: integer }
        enabled: { type: boolean }
        tags:
          type: object
          additionalProperties: { type: string }
        credential_kind: { type: string, description: A label — the secret itself is never returned. }
        group_name: { type: string, description: "Flat folder shown on `/targets`; empty: no folder." }
        position: { type: integer, format: int64, description: "Manual rank, tiebreaking only within the same state and folder." }
        last_probe_at: { $ref: "#/components/schemas/NullableServerTimestamp" }
        last_error: { type: [string, "null"] }
        error_kind:
          type: [string, "null"]
          enum: [down, config, null]
          description: "`down`: the device did not answer; `config`: our side (credentials, address, option)."
    NullableServerTimestamp:
      type: [string, "null"]
      description: UTC timestamp without suffix (`2026-08-31 20:15:00`), or `null`.
    TargetPayload:
      type: object
      required: [name, address, kind]
      properties:
        name: { type: string, maxLength: 200 }
        address: { type: string, maxLength: 253 }
        kind: { type: string }
        profile_id: { type: [string, "null"], description: "Absent on `PUT` keeps the detected profile; `\"\"` clears it." }
        parent_id: { type: [integer, "null"], format: int64 }
        via_agent: { type: [integer, "null"], format: int64, description: "Absent on `PUT` keeps the relay; `null` means direct probing." }
        interval_secs: { type: integer, minimum: 10, default: 60 }
        enabled: { type: boolean, default: true }
        tags:
          type: object
          description: Type options go here under their key.
          additionalProperties: { type: string }
        group_name: { type: string, description: "Flat folder shown on `/targets`. Absent or empty: no folder." }
        credential:
          $ref: "#/components/schemas/CredentialInput"
          description: 'Absent on `PUT` keeps the stored secret; `{"type": "none"}` clears it.'
    CredentialInput:
      description: A device credential. Accepted on input only — never returned.
      writeOnly: true
      oneOf:
        - type: object
          required: [type]
          properties:
            type: { const: none }
        - type: object
          required: [type, community]
          properties:
            type: { const: snmp_community }
            community: { type: string }
        - type: object
          required: [type, username]
          properties:
            type: { const: snmp_v3 }
            username: { type: string }
            auth:
              type: [object, "null"]
              required: [protocol, passphrase]
              properties:
                protocol: { type: string, enum: [md5, sha1, sha224, sha256, sha384, sha512] }
                passphrase: { type: string }
            privacy:
              type: [object, "null"]
              required: [protocol, passphrase]
              properties:
                protocol: { type: string, enum: [des, aes128, aes192, aes256] }
                passphrase: { type: string }
            context: { type: [string, "null"] }
        - type: object
          required: [type]
          description: "`token`, or the two halves `token_id` and `secret` that the server joins into `user@realm!name=secret` for Proxmox VE and PBS."
          properties:
            type: { const: api_token }
            token: { type: string }
            token_id: { type: string }
            secret: { type: string }
        - type: object
          required: [type, username, password]
          properties:
            type: { const: username_password }
            username: { type: string }
            password: { type: string }
...
    Severity:
      type: string
      enum: [info, warning, critical]
      description: Internal severity. The UI words are info, advisory (warning) and warning (critical).
    Phase:
      type: string
      enum: [ok, pending, firing, resolved]
    EffectivePhase:
      type: string
      enum: [ok, pending, firing, resolved, suppressed]
    ActiveAlert:
      type: object
      required: [fingerprint, rule_uid, rule_name, severity, series_key, labels, phase, effective_phase, suppressed, silenced, learning, acked, notify_count]
      properties:
        fingerprint: { type: string }
        rule_uid: { type: string }
        rule_name: { type: string, description: Empty if the rule was deleted meanwhile. }
        severity: { $ref: "#/components/schemas/Severity" }
        target_id: { type: [integer, "null"], format: int64 }
        series_key: { type: string }
        labels:
          type: object
          additionalProperties: { type: string }
        phase: { $ref: "#/components/schemas/Phase" }
        effective_phase: { $ref: "#/components/schemas/EffectivePhase" }
        suppressed: { type: boolean }
        suppressed_by: { type: [integer, "null"], format: int64, description: The unreachable parent device. }
        silenced: { type: boolean }
        learning: { type: boolean, description: Baseline rules stay silent for 14 days. }
        acked: { type: boolean }
        acked_until: { type: [string, "null"], format: date-time }
        acked_by: { type: [string, "null"], description: "An account, or `token:<name>`." }
        ack_note: { type: [string, "null"] }
        value: { type: [number, "null"] }
        score: { type: [number, "null"] }
        condition_since: { type: [string, "null"], format: date-time }
        firing_since: { type: [string, "null"], format: date-time }
        last_eval_at: { type: [string, "null"], format: date-time }
        last_notified_at: { type: [string, "null"], format: date-time }
        notify_count: { type: integer }
...
    BackupRestoreRequest:
      type: object
      required: [bundle, passphrase]
      properties:
        bundle: { type: object, additionalProperties: true }
        passphrase: { type: string, writeOnly: true }
        apply: { type: boolean, default: false, description: "`false`: dry run inside a transaction rolled back." }
    RestoreReport:
      type: object
      properties:
        applied: { type: boolean }
        created_at: { type: string }
        source_version: { type: string }
        sections:
          type: array
          items:
            type: object
            properties:
              section: { type: string }
              created: { type: integer }
              updated: { type: integer }
              skipped: { type: integer }
              notes:
                type: array
                items: { type: string }
        created: { type: integer }
        updated: { type: integer }
        skipped: { type: integer }
        warnings:
          type: array
          items: { type: string }

    # ------------------------------------------ panneaux (premier niveau seulement)
...
    McpEndpointInfo:
      type: object
      required: [error]
      properties:
        error: { type: string }
        protocolVersion: { type: string }
        transport: { type: string, const: streamable-http }
        streaming: { type: boolean }
        docs: { type: string }
    JsonRpcRequest:
      type: object
      required: [jsonrpc, method]
      description: A request (with `id`) or a notification (without). Batches are refused.
      properties:
        jsonrpc: { type: string, const: "2.0" }
        id: { type: [string, integer] }
        method:
          type: string
          examples: [server/discover, initialize, tools/list, tools/call, ping]
        params: { type: object, additionalProperties: true }
    JsonRpcResponse:
      type: object
      required: [jsonrpc]
      properties:
        jsonrpc: { type: string, const: "2.0" }
        id: { type: [string, integer, "null"] }
        result: { type: object, additionalProperties: true }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: integer }
            message: { type: string }
            data: {}
...